updated test for karavictl
