--- cmd/karavictl/cmd/storage_list_grpc_test.go
@@ -1,140 +1,0 @@
-// Copyright © 2021-2023 Dell Inc., or its subsidiaries. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package cmd
-
-import (
-	"bytes"
-	"context"
-	"encoding/json"
-	"errors"
-	"io"
-	"io/ioutil"
-	mockStorage "karavi-authorization/internal/storage-service/mocks"
-	"karavi-authorization/pb"
-	"os"
-	"strings"
-	"testing"
-
-	"google.golang.org/grpc"
-)
-
-func TestStorageListGrpc(t *testing.T) {
-	afterFn := func() {
-		CreateStorageServiceClient = createStorageServiceClient
-		JSONOutput = jsonOutput
-		osExit = os.Exit
-	}
-
-	t.Run("it requests list of storage", func(t *testing.T) {
-		defer afterFn()
-
-		listStorageFn := func(context.Context, *pb.StorageListRequest, ...grpc.CallOption) (*pb.StorageListResponse, error) {
-			storage := `{"powerflex":{"11e4e7d35817bd0f":{"User":"admin","Password":"test","Endpoint":"https://10.0.0.1","Insecure":false}}}`
-			return &pb.StorageListResponse{Storage: []byte(storage)}, nil
-		}
-
-		CreateStorageServiceClient = func(_ string, _ bool) (pb.StorageServiceClient, io.Closer, error) {
-			return &mockStorage.FakeStorageServiceClient{ListStorageFn: listStorageFn}, ioutil.NopCloser(nil), nil
-		}
-		osExit = func(code int) {
-		}
-
-		var gotOutput bytes.Buffer
-
-		cmd := NewRootCmd()
-		cmd.SetOutput(&gotOutput)
-		cmd.SetArgs([]string{"storage", "list", "--addr", "https://storage-service.com", "--insecure"})
-		cmd.Execute()
-
-		got := strings.ReplaceAll(gotOutput.String(), "\n", "")
-		got = strings.ReplaceAll(got, " ", "")
-
-		want := `{"powerflex":{"11e4e7d35817bd0f":{"Endpoint":"https://10.0.0.1","Insecure":false,"Password":"(omitted)","User":"admin"}}}`
-		if want != got {
-			t.Errorf("want %s, got \n%s", want, got)
-		}
-	})
-	t.Run("it requires a valid role server connection", func(t *testing.T) {
-		defer afterFn()
-		CreateStorageServiceClient = func(_ string, _ bool) (pb.StorageServiceClient, io.Closer, error) {
-			return nil, ioutil.NopCloser(nil), errors.New("test error")
-		}
-		var gotCode int
-		done := make(chan struct{})
-		osExit = func(code int) {
-			gotCode = code
-			done <- struct{}{}
-			done <- struct{}{} // we can't let this function return
-		}
-		var gotOutput bytes.Buffer
-
-		cmd := NewRootCmd()
-		cmd.SetErr(&gotOutput)
-		cmd.SetArgs([]string{"storage", "list", "--addr", "https://storage-service.com", "--insecure"})
-		go cmd.Execute()
-		<-done
-
-		wantCode := 1
-		if gotCode != wantCode {
-			t.Errorf("got exit code %d, want %d", gotCode, wantCode)
-		}
-		var gotErr CommandError
-		if err := json.NewDecoder(&gotOutput).Decode(&gotErr); err != nil {
-			t.Fatal(err)
-		}
-		wantErrMsg := "test error"
-		if gotErr.ErrorMsg != wantErrMsg {
-			t.Errorf("got err %q, want %q", gotErr.ErrorMsg, wantErrMsg)
-		}
-	})
-	t.Run("it handles server errors", func(t *testing.T) {
-		defer afterFn()
-		CreateStorageServiceClient = func(_ string, _ bool) (pb.StorageServiceClient, io.Closer, error) {
-			return &mockStorage.FakeStorageServiceClient{
-				ListStorageFn: func(context.Context, *pb.StorageListRequest, ...grpc.CallOption) (*pb.StorageListResponse, error) {
-					return nil, errors.New("test error")
-				},
-			}, ioutil.NopCloser(nil), nil
-		}
-		var gotCode int
-		done := make(chan struct{})
-		osExit = func(code int) {
-			gotCode = code
-			done <- struct{}{}
-			done <- struct{}{} // we can't let this function return
-		}
-		var gotOutput bytes.Buffer
-
-		rootCmd := NewRootCmd()
-		rootCmd.SetErr(&gotOutput)
-		rootCmd.SetArgs([]string{"storage", "list", "--addr", "https://storage-service.com", "--insecure"})
-
-		go rootCmd.Execute()
-		<-done
-
-		wantCode := 1
-		if gotCode != wantCode {
-			t.Errorf("got exit code %d, want %d", gotCode, wantCode)
-		}
-		var gotErr CommandError
-		if err := json.NewDecoder(&gotOutput).Decode(&gotErr); err != nil {
-			t.Fatal(err)
-		}
-		wantErrMsg := "test error"
-		if gotErr.ErrorMsg != wantErrMsg {
-			t.Errorf("got err %q, want %q", gotErr.ErrorMsg, wantErrMsg)
-		}
-	})
-}
